modify test case
